Check for invalid inputs

+++ src/main/java/com/github/seqware/downloaders/GNOSDownloader.java
@@ -2,7 +2,7 @@ package com.github.seqware.downloaders;
 
 public class GNOSDownloader implements WorkflowFileDownloader {
 
-	private String downloadKey = "";
+	private String downloadKey = null;
 	
 		
 	public void setDownloadKey(String downloadKey)
@@ -20,6 +20,26 @@ public class GNOSDownloader implements WorkflowFileDownloader {
 	@Override
 	public String getDownloadCommandString(String downloadDir, String workflowName, String ... objectIDs) {
 
+		if (this.downloadKey == null || this.downloadKey.trim().length() == 0)
+		{
+			throw new RuntimeException("downloadKey cannot be null/empty!");
+		}
+		
+		if (objectIDs.length == 0 || objectIDs[0] == null || objectIDs[0].trim().length() == 0)
+		{
+			throw new RuntimeException("objectIDs is null/empty or the first element is null/empty!");
+		}
+		
+		if (workflowName == null || workflowName.trim().length() == 0)
+		{
+			throw new RuntimeException("workflowName is null/empty!");
+		}
+
+		if (downloadDir == null || downloadDir.trim().length() == 0)
+		{
+			throw new RuntimeException("downloadDir is null/empty!");
+		}
+		
 		String getFilesCommand = "( docker run --rm --name get_"+workflowName+" "
 				+ " -v "+this.downloadKey+":/gnos.key "
 			    + " -v "+downloadDir+"/:/downloads/:rw"
